Separate Workbaskets::WorkflowBaseController dependents from others into separate module for clarity
