--- app/views/workbaskets/schedule_export_to_cds/new.js.slim
@@ -1,3 +1,0 @@
-| $(".js-workbasket-schedule-export-to-cds-form-container").html("#{j render('workbaskets/schedule_export_to_cds/form')}");
-| MainMenuInteractions.openModal('schedule_export_to_cds_popup');
-| MainMenuInteractions.scheduleExportToCdsFormInit();
