+++ app/views/workbaskets/workflows/approves/new.html.slim
@@ -7,16 +7,14 @@
       = link_to "Measures", measures_url
 
     li aria-current="page"
-      | Cross-check and create measures
+      | Approve new measures
 
 header
   h1.heading-large
-    | Cross-check and create measures
+    | Approve new measures
 
   .view-workbasket-container
-    = render "workbaskets/#{workbasket.type}/workflow_screens_parts/notifications/block", screen_type: :cross_check
+    = render "workbaskets/#{workbasket.type}/workflow_screens_parts/notifications/block", partial_name: "workbaskets/#{workbasket.type}/workflow_screens_parts/notifications/approves/general"
     = render "workbaskets/#{workbasket.type}/workflow_screens_parts/workbasket_details"
-    = render "workbaskets/#{workbasket.type}/workflow_screens_parts/summary_of_configuration"
-
     = render "workbaskets/shared/steps/review_and_submit/measures", read_only: true, record_type: workbasket.type
-    = render "workbaskets/cross_checks/form"
+    = render "workbaskets/workflows/approves/form"
