+++ app/views/workbaskets/workflows/schedule_export_to_cds/new.js.slim
@@ -1,0 +1,3 @@
+| $(".js-workbasket-schedule-export-to-cds-form-container").html("#{j render('workbaskets/workflows/schedule_export_to_cds/form')}");
+| MainMenuInteractions.openModal('schedule_export_to_cds_popup');
+| MainMenuInteractions.scheduleExportToCdsFormInit();
